# Frontend/dividend-portfolio-manager/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Questrade Portfolio Tracker API
  description: |
    A comprehensive Node.js application for tracking your Questrade portfolio with automatic token refresh, 
    real-time data sync, and detailed dividend tracking.
    
    ## Features
    - 🔐 Automatic token refresh (weekly)
    - 📊 Real-time portfolio data sync
    - 💰 Detailed dividend tracking and analysis
    - 📈 Portfolio performance metrics
    - 🎯 Snap quote integration for current prices
    - 📅 Historical snapshots
    - 🔄 Automatic data synchronization during market hours
    
    ## Multi-Person Support
    This API supports multiple persons/users, each with their own Questrade account(s) and tokens.
    
    ## Authentication
    Each person requires a valid Questrade refresh token. The system automatically manages access tokens.
    
  version: "2.0.1"
  contact:
    name: Portfolio Manager Support
    url: https://github.com/yourusername/portfolio-manager
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000/api
    description: Development server
  - url: https://your-production-domain.com/api
    description: Production server
...
paths:
...
  /portfolio/summary:
    get:
      tags: ["Portfolio"]
      summary: Get portfolio summary
      parameters:
        - $ref: '#/components/parameters/ViewMode'
        - $ref: '#/components/parameters/PersonNameQuery'
        - $ref: '#/components/parameters/AccountId'
        - name: aggregate
          in: query
          schema:
            type: boolean
            default: false
          description: Whether to aggregate positions
      responses:
        '200':
          description: Portfolio summary
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PortfolioSummary'
  /portfolio/positions:
    get:
      tags: ["Portfolio"]
      summary: Get portfolio positions
      parameters:
        - $ref: '#/components/parameters/ViewMode'
        - $ref: '#/components/parameters/PersonNameQuery'
        - $ref: '#/components/parameters/AccountId'
        - name: aggregate
          in: query
          schema:
            type: boolean
            default: false
          description: Whether to aggregate positions
      responses:
        '200':
          description: Portfolio positions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Position'
                      meta:
                        $ref: '#/components/schemas/PositionsMeta'
...
  /portfolio/cash-balances:
    get:
      tags: ["Portfolio"]
      summary: Get cash balances
      parameters:
        - $ref: '#/components/parameters/ViewMode'
        - $ref: '#/components/parameters/PersonNameQuery'
        - $ref: '#/components/parameters/AccountId'
        - name: currency
          in: query
          schema:
            type: string
            enum: [CAD, USD]
          description: Filter by currency
      responses:
        '200':
          description: Cash balances
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CashBalanceData'
...
components:
  parameters:
...
    PersonNameQuery:
      name: personName
      in: query
      schema:
        type: string
      example: "John Doe"
    AccountId:
      name: accountId
      in: query
      schema:
        type: string
      example: "53413547"
    ViewMode:
      name: viewMode
      in: query
      schema:
        type: string
        enum: [all, person, account]
        default: all
      description: View mode for filtering data
...
  schemas:
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
        timestamp:
          type: string
          format: date-time
...
    CurrencyBalance:
      type: object
      properties:
        currency:
          type: string
        cash:
          type: number
        marketValue:
          type: number
        totalEquity:
          type: number
        buyingPower:
          type: number
        maintenanceExcess:
          type: number
        isRealTime:
          type: boolean
...
    PortfolioSummary:
      type: object
      properties:
        viewMode:
          type: string
          enum: [all, person, account]
        personName:
          type: string
        accountId:
          type: string
        currency:
          type: string
        totalInvestment:
          type: number
          example: 75000.00
        currentValue:
          type: number
          example: 82500.00
        totalReturnValue:
          type: number
          example: 8500.00
        totalReturnPercent:
          type: number
          example: 11.33
        unrealizedPnl:
          type: number
          example: 7500.00
        totalDividends:
          type: number
          example: 1000.00
        monthlyDividendIncome:
          type: number
          example: 125.50
        annualProjectedDividend:
          type: number
          example: 1506.00
        averageYieldPercent:
          type: number
          example: 1.83
        yieldOnCostPercent:
          type: number
          example: 2.01
        numberOfPositions:
          type: integer
          example: 15
        numberOfAccounts:
          type: integer
          example: 3
        numberOfDividendStocks:
          type: integer
          example: 8
        sectorAllocation:
          type: array
          items:
            $ref: '#/components/schemas/AllocationItem'
        currencyBreakdown:
          type: array
          items:
            $ref: '#/components/schemas/CurrencyAllocation'
        personBreakdown:
          type: array
          items:
            $ref: '#/components/schemas/PersonAllocation'
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountSummary'
        aggregationInfo:
          type: object
          properties:
            hasAggregatedPositions:
              type: boolean
            totalAggregatedSymbols:
              type: integer
    Position:
      type: object
      properties:
        symbol:
          type: string
          example: "AAPL"
        symbolId:
          type: integer
          example: 8049
        personName:
          type: string
        accountId:
          type: string
        openQuantity:
          type: number
          example: 100
        currentMarketValue:
          type: number
          example: 15000.00
        currentPrice:
          type: number
          example: 150.00
        averageEntryPrice:
          type: number
          example: 140.00
        totalCost:
          type: number
          example: 14000.00
        dayPnl:
          type: number
          example: 50.00
        openPnl:
          type: number
          example: 1000.00
        totalReturnPercent:
          type: number
          example: 7.14
        totalReturnValue:
          type: number
          example: 1000.00
        dividendPerShare:
          type: number
          example: 0.24
        dividendData:
          $ref: '#/components/schemas/DividendData'
        industrySector:
          type: string
          example: "Technology"
        currency:
          type: string
          example: "USD"
        isDividendStock:
          type: boolean
        isAggregated:
          type: boolean
        numberOfAccounts:
          type: integer
        individualPositions:
          type: array
          items:
            $ref: '#/components/schemas/IndividualPosition'
        syncedAt:
          type: string
          format: date-time
    DividendData:
      type: object
      properties:
        totalReceived:
          type: number
          example: 250.00
        lastDividendAmount:
          type: number
          example: 24.00
        lastDividendDate:
          type: string
          format: date-time
        dividendReturnPercent:
          type: number
          example: 1.79
        yieldOnCost:
          type: number
          example: 6.86
        dividendAdjustedCost:
          type: number
          example: 13750.00
        dividendAdjustedCostPerShare:
          type: number
          example: 137.50
        monthlyDividend:
          type: number
          example: 8.00
        monthlyDividendPerShare:
          type: number
          example: 0.08
        annualDividend:
          type: number
          example: 96.00
        annualDividendPerShare:
          type: number
          example: 0.96
        dividendFrequency:
          type: integer
          example: 4
    IndividualPosition:
      type: object
      properties:
        accountId:
          type: string
        accountName:
          type: string
        accountType:
          type: string
        shares:
          type: number
        avgCost:
          type: number
        marketValue:
          type: number
        totalCost:
          type: number
        openPnl:
          type: number
    PositionsMeta:
      type: object
      properties:
        viewMode:
          type: string
        personName:
          type: string
        accountId:
          type: string
        aggregated:
          type: boolean
        count:
          type: integer
        currencySummary:
          type: array
          items:
            $ref: '#/components/schemas/CurrencyAllocation'
    CashBalanceData:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/CashBalanceAccount'
        summary:
          type: object
          properties:
            totalAccounts:
              type: integer
            totalPersons:
              type: integer
            totalCAD:
              type: number
            totalUSD:
              type: number
        currencyBreakdown:
          type: array
          items:
            $ref: '#/components/schemas/CurrencyTotal'
        lastUpdated:
          type: string
          format: date-time
        viewMode:
          type: string
        personName:
          type: string
        accountId:
          type: string
    CashBalanceAccount:
      type: object
      properties:
        accountId:
          type: string
        accountName:
          type: string
        accountType:
          type: string
        personName:
          type: string
        cashBalance:
          type: number
        currency:
          type: string
        cashBalances:
          type: array
          items:
            $ref: '#/components/schemas/CurrencyBalance'
        lastUpdated:
          type: string
          format: date-time
        syncError:
          type: string
    CurrencyTotal:
      type: object
      properties:
        currency:
          type: string
        total:
          type: number
        percentage:
          type: number
...
    AllocationItem:
      type: object
      properties:
        sector:
          type: string
          example: "Technology"
        category:
          type: string
          example: "Stocks"
        value:
          type: number
          example: 25000.00
        percentage:
          type: number
          example: 30.30
    CurrencyAllocation:
      type: object
      properties:
        currency:
          type: string
          example: "CAD"
        value:
          type: number
          example: 50000.00
        percentage:
          type: number
          example: 60.61
    PersonAllocation:
      type: object
      properties:
        personName:
          type: string
        value:
          type: number
        percentage:
          type: number
        numberOfPositions:
          type: integer
    AccountSummary:
      type: object
      properties:
        accountId:
          type: string
        accountName:
          type: string
        accountType:
          type: string
        currency:
          type: string
        totalInvestment:
          type: number
        currentValue:
          type: number
        unrealizedPnl:
          type: number
        cashBalance:
          type: number
        numberOfPositions:
          type: integer
        returnPercent:
          type: number
        lastUpdated:
          type: string
          format: date-time
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (if implemented)

# Security (if you implement API key authentication)
# security:
#   - ApiKeyAuth: []